requirements.txt
beautifulsoup4==4.10.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cssselect==1.1.0
cssutils==2.4.0
idna==3.3
lxml==4.8.0
premailer==3.10.0
python-dotenv==0.19.2
requests==2.27.1
soupsieve==2.3.1
urllib3==1.26.8
yagmail==0.15.277
